agent_odonto_4/requirements.txt
# Principais bibliotecas
faiss-cpu
openai
python-dotenv
langchain
langchainhub
langchain-core
langchain-text-splitters
langchain-experimental
langchain-community
langchain-huggingface
chromadb
pydantic
pandas
streamlit
sentence-transformers
PyPDF
python-docx
# Dependências auxiliares
aiohttp
altair
anyio
attrs
bcrypt
blinker
certifi==2024.8.30
charset-normalizer
click
decorator
fastapi
fsspec
gitpython
grpcio
huggingface-hub
idna
ipykernel
ipython
jinja2
joblib
markdown-it-py
matplotlib-inline
numpy
pillow
protobuf
pytz
regex
requests
rich
scikit-learn
scipy
sqlalchemy
tqdm
transformers
uvicorn
websockets
